+++ src/PhpSpreadsheet/Calculation/Calculation.php
@@ -132,6 +132,7 @@ class Calculation extends CalculationLocale
      */
     private CyclicReferenceStack $cyclicReferenceStack;
 
+    /** @var mixed[] */
     private array $cellStack = [];
 
     /**
@@ -496,6 +497,7 @@ class Calculation extends CalculationLocale
             if ($cellAddress === null) {
                 throw new Exception('null cellAddress in calculateCellValue');
             }
+            /** @var array{sheet: string, cell: string} $cellAddress */
             $testSheet = $this->spreadsheet->getSheetByName($cellAddress['sheet']);
             if ($testSheet === null) {
                 throw new Exception('worksheet not found in calculateCellValue');
@@ -509,6 +511,7 @@ class Calculation extends CalculationLocale
                 $sheetName = $cellAddress['sheet'] ?? null;
                 $testSheet = is_string($sheetName) ? $this->spreadsheet->getSheetByName($sheetName) : null;
                 if ($testSheet !== null && array_key_exists('cell', $cellAddress)) {
+                    /** @var array{cell: string} $cellAddress */
                     $testSheet->getCell($cellAddress['cell']);
                 }
             }
@@ -537,6 +540,8 @@ class Calculation extends CalculationLocale
      * Validate and parse a formula string.
      *
      * @param string $formula Formula to parse
+     *
+     * @return array<mixed>|bool
      */
     public function parseFormula(string $formula): array|bool
     {
@@ -705,11 +710,11 @@ class Calculation extends CalculationLocale
      *
      * @param mixed $operand1 First matrix operand
      *
-     * @param-out array $operand1
+     * @param-out mixed[] $operand1
      *
      * @param mixed $operand2 Second matrix operand
      *
-     * @param-out array $operand2
+     * @param-out mixed[] $operand2
      *
      * @param int $resize Flag indicating whether the matrices should be resized to match
      *                                        and (if so), whether the smaller dimension should grow or the
@@ -954,14 +959,14 @@ class Calculation extends CalculationLocale
         return null;
     }
 
+    private const MATRIX_REPLACE_FROM = [self::FORMULA_OPEN_MATRIX_BRACE, ';', self::FORMULA_CLOSE_MATRIX_BRACE];
+    private const MATRIX_REPLACE_TO = ['MKMATRIX(MKMATRIX(', '),MKMATRIX(', '))'];
+
     /**
      * @return false|string False indicates an error
      */
     private function convertMatrixReferences(string $formula): false|string
     {
-        static $matrixReplaceFrom = [self::FORMULA_OPEN_MATRIX_BRACE, ';', self::FORMULA_CLOSE_MATRIX_BRACE];
-        static $matrixReplaceTo = ['MKMATRIX(MKMATRIX(', '),MKMATRIX(', '))'];
-
         //    Convert any Excel matrix references to the MKMATRIX() function
         if (str_contains($formula, self::FORMULA_OPEN_MATRIX_BRACE)) {
             //    If there is the possibility of braces within a quoted string, then we don't treat those as matrix indicators
@@ -978,7 +983,7 @@ class Calculation extends CalculationLocale
                     if ($notWithinQuotes === true) {
                         $openCount += substr_count($value, self::FORMULA_OPEN_MATRIX_BRACE);
                         $closeCount += substr_count($value, self::FORMULA_CLOSE_MATRIX_BRACE);
-                        $value = str_replace($matrixReplaceFrom, $matrixReplaceTo, $value);
+                        $value = str_replace(self::MATRIX_REPLACE_FROM, self::MATRIX_REPLACE_TO, $value);
                     }
                 }
                 unset($value);
@@ -988,7 +993,7 @@ class Calculation extends CalculationLocale
                 //    If there's no quoted strings, then we do a simple count/replace
                 $openCount = substr_count($formula, self::FORMULA_OPEN_MATRIX_BRACE);
                 $closeCount = substr_count($formula, self::FORMULA_CLOSE_MATRIX_BRACE);
-                $formula = str_replace($matrixReplaceFrom, $matrixReplaceTo, $formula);
+                $formula = str_replace(self::MATRIX_REPLACE_FROM, self::MATRIX_REPLACE_TO, $formula);
             }
             //    Trap for mismatched braces and trigger an appropriate error
             if ($openCount < $closeCount) {
@@ -1120,7 +1125,7 @@ class Calculation extends CalculationLocale
                 // call or a parenthesis
                 $this->branchPruner->decrementDepth();
 
-                if (is_array($d) && preg_match('/^' . self::CALCULATION_REGEXP_FUNCTION . '$/miu', $d['value'], $matches)) {
+                if (is_array($d) && preg_match('/^' . self::CALCULATION_REGEXP_FUNCTION . '$/miu', StringHelper::convertToString($d['value']), $matches)) {
                     //    Did this parenthesis just close a function?
                     try {
                         $this->branchPruner->closingBrace($d['value']);
@@ -1184,6 +1189,7 @@ class Calculation extends CalculationLocale
                         }
                     }
                     if ($argumentCountError) {
+                        /** @var int $argumentCount */
                         return $this->raiseFormulaError("Formula Error: Wrong number of arguments for $functionName() function: $argumentCount given, " . $expectedArgumentCountString . ' expected');
                     }
                 }
@@ -1205,7 +1211,9 @@ class Calculation extends CalculationLocale
                 }
                 // make sure there was a function
                 $d = $stack->last(2);
-                if (!preg_match('/^' . self::CALCULATION_REGEXP_FUNCTION . '$/miu', $d['value'] ?? '', $matches)) {
+                /** @var string */
+                $temp = $d['value'] ?? '';
+                if (!preg_match('/^' . self::CALCULATION_REGEXP_FUNCTION . '$/miu', $temp, $matches)) {
                     // Can we inject a dummy function at this point so that the braces at least have some context
                     //     because at least the braces are paired up (at this stage in the formula)
                     // MS Excel allows this if the content is cell references; but doesn't allow actual values,
@@ -1273,6 +1281,7 @@ class Calculation extends CalculationLocale
                                 // Do we have chained range operators?
                                 $rangeStartCellRef = $output[count($output) - 2]['value'] ?? '';
                             }
+                            /** @var string $rangeStartCellRef */
                             preg_match('/^' . self::CALCULATION_REGEXP_CELLREF . '$/miu', $rangeStartCellRef, $rangeStartMatches);
                             if (array_key_exists(2, $rangeStartMatches)) {
                                 if ($rangeStartMatches[2] > '') {
@@ -1287,6 +1296,7 @@ class Calculation extends CalculationLocale
                                 // Do we have chained range operators?
                                 $rangeStartCellRef = $output[count($output) - 2]['value'] ?? '';
                             }
+                            /** @var string $rangeStartCellRef */
                             preg_match('/^' . self::CALCULATION_REGEXP_CELLREF . '$/miu', $rangeStartCellRef, $rangeStartMatches);
                             if (isset($rangeStartMatches[2]) && $rangeStartMatches[2] !== $matches[2]) {
                                 return $this->raiseFormulaError('3D Range references are not yet supported');
@@ -1547,6 +1557,8 @@ class Calculation extends CalculationLocale
     private static int $matchIndex10 = 10;
 
     /**
+     * @param array<mixed>|false $tokens
+     *
      * @return array<int, mixed>|false|string
      */
     private function processTokenStack(false|array $tokens, ?string $cellID = null, ?Cell $cell = null)
@@ -1569,14 +1581,17 @@ class Calculation extends CalculationLocale
         $branchStore = [];
         //    Loop through each token in turn
         foreach ($tokens as $tokenIdx => $tokenData) {
+            /** @var mixed[] $tokenData */
             $this->processingAnchorArray = false;
-            if ($tokenData['type'] === 'Cell Reference' && isset($tokens[$tokenIdx + 1]) && $tokens[$tokenIdx + 1]['type'] === 'Operand Count for Function ANCHORARRAY()') {
+            if ($tokenData['type'] === 'Cell Reference' && isset($tokens[$tokenIdx + 1]) && $tokens[$tokenIdx + 1]['type'] === 'Operand Count for Function ANCHORARRAY()') { //* @phpstan-ignore-line
                 $this->processingAnchorArray = true;
             }
             $token = $tokenData['value'];
             // Branch pruning: skip useless resolutions
+            /** @var ?string */
             $storeKey = $tokenData['storeKey'] ?? null;
             if ($this->branchPruningEnabled && isset($tokenData['onlyIf'])) {
+                /** @var string */
                 $onlyIfStoreKey = $tokenData['onlyIf'];
                 $storeValue = $branchStore[$onlyIfStoreKey] ?? null;
                 $storeValueAsBool = ($storeValue === null)
@@ -1591,7 +1606,9 @@ class Calculation extends CalculationLocale
                     && (!$storeValueAsBool || Information\ErrorValue::isError($storeValue) || ($storeValue === 'Pruned branch'))
                 ) {
                     // If branching value is not true, we don't need to compute
+                    /** @var string $onlyIfStoreKey */
                     if (!isset($fakedForBranchPruning['onlyIf-' . $onlyIfStoreKey])) {
+                        /** @var string $token */
                         $stack->push('Value', 'Pruned branch (only if ' . $onlyIfStoreKey . ') ' . $token);
                         $fakedForBranchPruning['onlyIf-' . $onlyIfStoreKey] = true;
                     }
@@ -1609,6 +1626,7 @@ class Calculation extends CalculationLocale
             }
 
             if ($this->branchPruningEnabled && isset($tokenData['onlyIfNot'])) {
+                /** @var string */
                 $onlyIfNotStoreKey = $tokenData['onlyIfNot'];
                 $storeValue = $branchStore[$onlyIfNotStoreKey] ?? null;
                 $storeValueAsBool = ($storeValue === null)
@@ -1624,6 +1642,7 @@ class Calculation extends CalculationLocale
                 ) {
                     // If branching value is true, we don't need to compute
                     if (!isset($fakedForBranchPruning['onlyIfNot-' . $onlyIfNotStoreKey])) {
+                        /** @var string $token */
                         $stack->push('Value', 'Pruned branch (only if not ' . $onlyIfNotStoreKey . ') ' . $token);
                         $fakedForBranchPruning['onlyIfNot-' . $onlyIfNotStoreKey] = true;
                     }
@@ -1706,19 +1725,22 @@ class Calculation extends CalculationLocale
                     // Binary Operators
                     case ':': // Range
                         if ($operand1Data['type'] === 'Defined Name') {
+                            /** @var array{reference: string} $operand1Data */
                             if (preg_match('/$' . self::CALCULATION_REGEXP_DEFINEDNAME . '^/mui', $operand1Data['reference']) !== false && $this->spreadsheet !== null) {
+                                /** @var string[] $operand1Data */
                                 $definedName = $this->spreadsheet->getNamedRange($operand1Data['reference']);
                                 if ($definedName !== null) {
                                     $operand1Data['reference'] = $operand1Data['value'] = str_replace('$', '', $definedName->getValue());
                                 }
                             }
                         }
+                        /** @var array{reference?: ?string} $operand1Data */
                         if (str_contains($operand1Data['reference'] ?? '', '!')) {
                             [$sheet1, $operand1Data['reference']] = Worksheet::extractSheetTitle($operand1Data['reference'], true, true);
                         } else {
                             $sheet1 = ($pCellWorksheet !== null) ? $pCellWorksheet->getTitle() : '';
                         }
-                        $sheet1 ??= '';
+                        //$sheet1 ??= ''; // phpstan level 10 says this is unneeded
 
                         /** @var string */
                         $op2ref = $operand2Data['reference'];
@@ -1728,6 +1750,7 @@ class Calculation extends CalculationLocale
                         }
 
                         if ($sheet1 === $sheet2) {
+                            /** @var array{reference: ?string, value: string|string[]} $operand1Data */
                             if ($operand1Data['reference'] === null && $cell !== null) {
                                 if (is_array($operand1Data['value'])) {
                                     $operand1Data['reference'] = $cell->getCoordinate();
@@ -1739,6 +1762,7 @@ class Calculation extends CalculationLocale
                                     $operand1Data['reference'] = $operand1Data['value'] . $cell->getRow();
                                 }
                             }
+                            /** @var array{reference: ?string, value: string|string[]} $operand2Data */
                             if ($operand2Data['reference'] === null && $cell !== null) {
                                 if (is_array($operand2Data['value'])) {
                                     $operand2Data['reference'] = $cell->getCoordinate();
@@ -1813,7 +1837,9 @@ class Calculation extends CalculationLocale
 
                             for ($row = 0; $row < $rows; ++$row) {
                                 for ($column = 0; $column < $columns; ++$column) {
+                                    /** @var mixed[][] $operand1 */
                                     $op1x = self::boolToString($operand1[$row][$column]);
+                                    /** @var mixed[][] $operand2 */
                                     $op2x = self::boolToString($operand2[$row][$column]);
                                     if (Information\ErrorValue::isError($op1x)) {
                                         // no need to do anything
@@ -1898,8 +1924,11 @@ class Calculation extends CalculationLocale
                     [$rows, $columns] = self::checkMatrixOperands($result, $operand2, 0);
                     for ($row = 0; $row < $rows; ++$row) {
                         for ($column = 0; $column < $columns; ++$column) {
+                            /** @var mixed[][] $result */
                             if (self::isNumericOrBool($result[$row][$column])) {
-                                $result[$row][$column] *= $multiplier;
+                                /** @var float|int|numeric-string */
+                                $temp = $result[$row][$column];
+                                $result[$row][$column] = $temp * $multiplier;
                             } else {
                                 $result[$row][$column] = self::makeError($result[$row][$column]);
                             }
@@ -1914,7 +1943,7 @@ class Calculation extends CalculationLocale
                 } else {
                     $this->executeNumericBinaryOperation($multiplier, $arg, '*', $stack);
                 }
-            } elseif (preg_match('/^' . self::CALCULATION_REGEXP_CELLREF . '$/i', $token ?? '', $matches)) {
+            } elseif (preg_match('/^' . self::CALCULATION_REGEXP_CELLREF . '$/i', StringHelper::convertToString($token ?? ''), $matches)) {
                 $cellRef = null;
 
                 /* Phpstan says matches[8/9/10] is never set,
@@ -2005,7 +2034,7 @@ class Calculation extends CalculationLocale
                 if (isset($storeKey)) {
                     $branchStore[$storeKey] = $cellValue;
                 }
-            } elseif (preg_match('/^' . self::CALCULATION_REGEXP_FUNCTION . '$/miu', $token ?? '', $matches)) {
+            } elseif (preg_match('/^' . self::CALCULATION_REGEXP_FUNCTION . '$/miu', StringHelper::convertToString($token ?? ''), $matches)) {
                 // if the token is a function, pop arguments off the stack, hand them to the function, and push the result back on
                 if ($cell !== null && $pCellParent !== null) {
                     $cell->attach($pCellParent);
@@ -2131,14 +2160,16 @@ class Calculation extends CalculationLocale
                 }
             } else {
                 // if the token is a number, boolean, string or an Excel error, push it onto the stack
+                /** @var ?string $token */
                 if (isset(self::EXCEL_CONSTANTS[strtoupper($token ?? '')])) {
-                    $excelConstant = strtoupper($token);
+                    $excelConstant = strtoupper("$token");
                     $stack->push('Constant Value', self::EXCEL_CONSTANTS[$excelConstant]);
                     if (isset($storeKey)) {
                         $branchStore[$storeKey] = self::EXCEL_CONSTANTS[$excelConstant];
                     }
                     $this->debugLog->writeDebugLog('Evaluating Constant %s as %s', $excelConstant, $this->showTypeDetails(self::EXCEL_CONSTANTS[$excelConstant]));
-                } elseif ((is_numeric($token)) || ($token === null) || (is_bool($token)) || ($token == '') || ($token[0] == self::FORMULA_STRING_QUOTE) || ($token[0] == '#')) {
+                } elseif ((is_numeric($token)) || ($token === null) || (is_bool($token)) || ($token == '') || ($token[0] == self::FORMULA_STRING_QUOTE) || ($token[0] == '#')) { //* @phpstan-ignore-line
+                    /** @var array{type: string, reference: ?string} $tokenData */
                     $stack->push($tokenData['type'], $token, $tokenData['reference']);
                     if (isset($storeKey)) {
                         $branchStore[$storeKey] = $token;
@@ -2464,6 +2495,7 @@ class Calculation extends CalculationLocale
     public function extractCellRange(string &$range = 'A1', ?Worksheet $worksheet = null, bool $resetLog = true): array
     {
         // Return value
+        /** @var mixed[][] */
         $returnValue = [];
 
         if ($worksheet !== null) {
@@ -2557,6 +2589,7 @@ class Calculation extends CalculationLocale
             if (!isset($aReferences[1])) {
                 //    Single cell (or single column or row) in range
                 [$currentCol, $currentRow] = Coordinate::coordinateFromString($aReferences[0]);
+                /** @var mixed[][] $returnValue */
                 if ($worksheet !== null && $worksheet->cellExists($aReferences[0])) {
                     $returnValue[$currentRow][$currentCol] = $worksheet->getCell($aReferences[0])->getCalculatedValue($resetLog);
                 } else {