+++ src/PhpSpreadsheet/Worksheet/Worksheet.php
@@ -3,6 +3,7 @@
 namespace PhpOffice\PhpSpreadsheet\Worksheet;
 
 use ArrayObject;
+use Composer\Pcre\Preg;
 use Generator;
 use PhpOffice\PhpSpreadsheet\Calculation\Calculation;
 use PhpOffice\PhpSpreadsheet\Calculation\Functions;
@@ -287,12 +288,16 @@ class Worksheet
 
     /**
      * Hyperlinks. Indexed by cell coordinate, e.g. 'A1'.
+     *
+     * @var Hyperlink[]
      */
     private array $hyperlinkCollection = [];
 
     /**
      * Data validation objects. Indexed by cell coordinate, e.g. 'A1'.
      * Index can include ranges, and multiple cells/ranges.
+     *
+     * @var DataValidation[]
      */
     private array $dataValidationCollection = [];
 
@@ -389,6 +394,8 @@ class Worksheet
 
     /**
      * Get array of invalid characters for sheet title.
+     *
+     * @return string[]
      */
     public static function getInvalidCharacters(): array
     {
@@ -729,6 +736,7 @@ class Worksheet
                         $filterAdjustment = false;
                         if (!empty($autoFilterIndentRanges)) {
                             foreach ($autoFilterIndentRanges as $autoFilterFirstRowRange) {
+                                /** @var string $autoFilterFirstRowRange */
                                 if ($cell->isInRange($autoFilterFirstRowRange)) {
                                     $filterAdjustment = true;
 
@@ -1101,7 +1109,7 @@ class Worksheet
     /**
      * Get highest worksheet column and highest row that have cell records.
      *
-     * @return array Highest column name and highest row number
+     * @return array{row: int, column: string} Highest column name and highest row number
      */
     public function getHighestRowAndColumn(): array
     {
@@ -1205,8 +1213,8 @@ class Worksheet
                 throw new Exception('Sheet not found for name: ' . $worksheetReference[0]);
             }
         } elseif (
-            !preg_match('/^' . Calculation::CALCULATION_REGEXP_CELLREF . '$/i', $coordinate)
-            && preg_match('/^' . Calculation::CALCULATION_REGEXP_DEFINEDNAME . '$/iu', $coordinate)
+            !Preg::isMatch('/^' . Calculation::CALCULATION_REGEXP_CELLREF . '$/i', $coordinate)
+            && Preg::isMatch('/^' . Calculation::CALCULATION_REGEXP_DEFINEDNAME . '$/iu', $coordinate)
         ) {
             // Named range?
             $namedRange = $this->validateNamedRange($coordinate, true);
@@ -1421,6 +1429,8 @@ class Worksheet
      *
      * @param Cell $cell
      *              The Cell for which the tables are retrieved
+     *
+     * @return mixed[]
      */
     public function getTablesWithStylesForCell(Cell $cell): array
     {
@@ -1458,7 +1468,7 @@ class Worksheet
     public function getConditionalStyles(string $coordinate, bool $firstOnly = true): array
     {
         $coordinate = strtoupper($coordinate);
-        if (preg_match('/[: ,]/', $coordinate) === 1) {
+        if (Preg::isMatch('/[: ,]/', $coordinate)) {
             return $this->conditionalStylesCollection[$coordinate] ?? [];
         }
 
@@ -1778,7 +1788,7 @@ class Worksheet
             $range .= ":{$range}";
         }
 
-        if (preg_match('/^([A-Z]+)(\d+):([A-Z]+)(\d+)$/', $range, $matches) !== 1) {
+        if (!Preg::isMatch('/^([A-Z]+)(\d+):([A-Z]+)(\d+)$/', $range, $matches)) {
             throw new Exception('Merge must be on a valid range of cells.');
         }
 
@@ -1861,6 +1871,11 @@ class Worksheet
         }
     }
 
+    /**
+     * @param mixed[] $leftCellValue
+     *
+     * @return mixed[]
+     */
     public function mergeCellBehaviour(Cell $cell, string $upperLeft, string $behaviour, array $leftCellValue): array
     {
         if ($cell->getCoordinate() !== $upperLeft) {
@@ -2392,6 +2407,42 @@ class Worksheet
         if ($row < 1) {
             throw new Exception('Rows to be deleted should at least start from row 1.');
         }
+        $startRow = $row;
+        $endRow = $startRow + $numberOfRows - 1;
+        $removeKeys = [];
+        $addKeys = [];
+        foreach ($this->mergeCells as $key => $value) {
+            if (
+                Preg::isMatch(
+                    '/^([a-z]{1,3})(\d+):([a-z]{1,3})(\d+)/i',
+                    $key,
+                    $matches
+                )
+            ) {
+                $startMergeInt = (int) $matches[2];
+                $endMergeInt = (int) $matches[4];
+                if ($startMergeInt >= $startRow) {
+                    if ($startMergeInt <= $endRow) {
+                        $removeKeys[] = $key;
+                    }
+                } elseif ($endMergeInt >= $startRow) {
+                    if ($endMergeInt <= $endRow) {
+                        $temp = $endMergeInt - 1;
+                        $removeKeys[] = $key;
+                        if ($temp !== $startMergeInt) {
+                            $temp3 = $matches[1] . $matches[2] . ':' . $matches[3] . $temp;
+                            $addKeys[] = $temp3;
+                        }
+                    }
+                }
+            }
+        }
+        foreach ($removeKeys as $key) {
+            unset($this->mergeCells[$key]);
+        }
+        foreach ($addKeys as $key) {
+            $this->mergeCells[$key] = $key;
+        }
 
         $holdRowDimensions = $this->removeRowDimensions($row, $numberOfRows);
         $highestRow = $this->getHighestDataRow();
@@ -2416,6 +2467,7 @@ class Worksheet
         return $this;
     }
 
+    /** @return RowDimension[] */
     private function removeRowDimensions(int $row, int $numberOfRows): array
     {
         $highRow = $row + $numberOfRows - 1;
@@ -2448,6 +2500,43 @@ class Worksheet
         if (is_numeric($column)) {
             throw new Exception('Column references should not be numeric.');
         }
+        $startColumnInt = Coordinate::columnIndexFromString($column);
+        $endColumnInt = $startColumnInt + $numberOfColumns - 1;
+        $removeKeys = [];
+        $addKeys = [];
+        foreach ($this->mergeCells as $key => $value) {
+            if (
+                Preg::isMatch(
+                    '/^([a-z]{1,3})(\d+):([a-z]{1,3})(\d+)/i',
+                    $key,
+                    $matches
+                )
+            ) {
+                $startMergeInt = Coordinate::columnIndexFromString($matches[1]);
+                $endMergeInt = Coordinate::columnIndexFromString($matches[3]);
+                if ($startMergeInt >= $startColumnInt) {
+                    if ($startMergeInt <= $endColumnInt) {
+                        $removeKeys[] = $key;
+                    }
+                } elseif ($endMergeInt >= $startColumnInt) {
+                    if ($endMergeInt <= $endColumnInt) {
+                        $temp = Coordinate::columnIndexFromString($matches[3]) - 1;
+                        $temp2 = Coordinate::stringFromColumnIndex($temp);
+                        $removeKeys[] = $key;
+                        if ($temp2 !== $matches[1]) {
+                            $temp3 = $matches[1] . $matches[2] . ':' . $temp2 . $matches[4];
+                            $addKeys[] = $temp3;
+                        }
+                    }
+                }
+            }
+        }
+        foreach ($removeKeys as $key) {
+            unset($this->mergeCells[$key]);
+        }
+        foreach ($addKeys as $key) {
+            $this->mergeCells[$key] = $key;
+        }
 
         $highestColumn = $this->getHighestDataColumn();
         $highestColumnIndex = Coordinate::columnIndexFromString($highestColumn);
@@ -2477,6 +2566,7 @@ class Worksheet
         return $this;
     }
 
+    /** @return ColumnDimension[] */
     private function removeColumnDimensions(int $pColumnIndex, int $numberOfColumns): array
     {
         $highCol = $pColumnIndex + $numberOfColumns - 1;
@@ -2809,7 +2899,7 @@ class Worksheet
     /**
      * Fill worksheet from values in array.
      *
-     * @param array $source Source array
+     * @param mixed[]|mixed[][] $source Source array
      * @param mixed $nullValue Value in source array that stands for blank cell
      * @param string $startCell Insert array starting from this cell address as the top left coordinate
      * @param bool $strictNullComparison Apply strict comparison when testing for null values in the array
@@ -2822,19 +2912,23 @@ class Worksheet
         if (!is_array(end($source))) {
             $source = [$source];
         }
+        /** @var mixed[][] $source */
 
         // start coordinate
         [$startColumn, $startRow] = Coordinate::coordinateFromString($startCell);
+        $startRow = (int) $startRow;
 
         // Loop through $source
         if ($strictNullComparison) {
             foreach ($source as $rowData) {
+                /** @var string */
                 $currentColumn = $startColumn;
                 foreach ($rowData as $cellValue) {
                     if ($cellValue !== $nullValue) {
-                        // Set cell value
+                        /** @var string $currentColumn */
                         $this->getCell($currentColumn . $startRow)->setValue($cellValue);
                     }
+                    /** @var string $currentColumn */
                     ++$currentColumn;
                 }
                 ++$startRow;
@@ -2844,9 +2938,10 @@ class Worksheet
                 $currentColumn = $startColumn;
                 foreach ($rowData as $cellValue) {
                     if ($cellValue != $nullValue) {
-                        // Set cell value
+                        /** @var string $currentColumn */
                         $this->getCell($currentColumn . $startRow)->setValue($cellValue);
                     }
+                    /** @var string $currentColumn */
                     ++$currentColumn;
                 }
                 ++$startRow;
@@ -2930,6 +3025,8 @@ class Worksheet
      *                             True - Return rows and columns indexed by their actual row and column IDs
      * @param bool $ignoreHidden False - Return values for rows/columns even if they are defined as hidden.
      *                            True - Don't return values for rows/columns that are defined as hidden.
+     *
+     * @return mixed[][]
      */
     public function rangesToArray(
         string $ranges,
@@ -2965,7 +3062,7 @@ class Worksheet
      * @param bool $ignoreHidden False - Return values for rows/columns even if they are defined as hidden.
      *                            True - Don't return values for rows/columns that are defined as hidden.
      *
-     * @return Generator<array>
+     * @return Generator<array<mixed>>
      */
     public function rangeToArrayYieldRows(
         string $range,
@@ -3049,6 +3146,8 @@ class Worksheet
      * @param bool $ignoreHidden False - Return values for rows/columns even if they are defined as hidden.
      *                             True - Don't return values for rows/columns that are defined as hidden.
      * @param array<string, bool> $hiddenColumns
+     *
+     * @return mixed[]
      */
     private function buildNullRow(
         mixed $nullValue,
@@ -3061,8 +3160,9 @@ class Worksheet
         $nullRow = [];
         $c = -1;
         for ($col = $minCol; $col !== $maxCol; ++$col) {
+            /** @var string $col */
             if ($ignoreHidden === true && $this->columnDimensionExists($col) && $this->getColumnDimension($col)->getVisible() === false) {
-                $hiddenColumns[$col] = true; // @phpstan-ignore-line
+                $hiddenColumns[$col] = true;
             } else {
                 $columnRef = $returnCellRef ? $col : ++$c;
                 $nullRow[$columnRef] = $nullValue;
@@ -3230,6 +3330,7 @@ class Worksheet
         } else {
             $this->cachedHighestColumn = $highestColumn;
         }
+        /** @var int $highestRow */
         $this->cachedHighestRow = $highestRow;
 
         // Return
@@ -3436,7 +3537,7 @@ class Worksheet
         $collectionCells = [];
         $collectionRanges = [];
         foreach ($this->dataValidationCollection as $key => $dataValidation) {
-            if (preg_match('/[: ]/', $key) === 1) {
+            if (Preg::isMatch('/[: ]/', $key)) {
                 $collectionRanges[$key] = $dataValidation;
             } else {
                 $collectionCells[$key] = $dataValidation;
@@ -3704,7 +3805,7 @@ class Worksheet
 
     public static function nameRequiresQuotes(string $sheetName): bool
     {
-        return preg_match(self::SHEET_NAME_REQUIRES_NO_QUOTES, $sheetName) !== 1;
+        return !Preg::isMatch(self::SHEET_NAME_REQUIRES_NO_QUOTES, $sheetName);
     }
 
     public function isRowVisible(int $row): bool
@@ -3837,7 +3938,7 @@ class Worksheet
             $keys = $this->cellCollection->getCoordinates();
             foreach ($keys as $key) {
                 if ($this->getCell($key)->getDataType() === DataType::TYPE_FORMULA) {
-                    if (preg_match(self::FUNCTION_LIKE_GROUPBY, $this->getCell($key)->getValueString()) !== 1) {
+                    if (!Preg::isMatch(self::FUNCTION_LIKE_GROUPBY, $this->getCell($key)->getValueString())) {
                         $this->getCell($key)->getCalculatedValue();
                     }
                 }