+++ tests/PhpSpreadsheetTests/Calculation/Functions/DateTime/Days360Test.php
@@ -12,18 +12,19 @@ use PhpOffice\PhpSpreadsheet\Calculation\DateTimeExcel\Days360;
 use PhpOffice\PhpSpreadsheet\Calculation\Exception as CalculationException;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
 use PhpOffice\PhpSpreadsheetTests\Calculation\Functions\FormulaArguments;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 class Days360Test extends TestCase
 {
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerDAYS360')]
+    #[DataProvider('providerDAYS360')]
     public function testDirectCallToDAYS360(mixed $expectedResult, mixed ...$args): void
     {
         $result = Days360::between(...$args);
         self::assertSame($expectedResult, $result);
     }
 
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerDAYS360')]
+    #[DataProvider('providerDAYS360')]
     public function testDAYS360AsFormula(mixed $expectedResult, mixed ...$args): void
     {
         $arguments = new FormulaArguments(...$args);
@@ -35,7 +36,7 @@ class Days360Test extends TestCase
         self::assertSame($expectedResult, $result);
     }
 
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerDAYS360')]
+    #[DataProvider('providerDAYS360')]
     public function testDAYS360InWorksheet(mixed $expectedResult, mixed ...$args): void
     {
         $arguments = new FormulaArguments(...$args);
@@ -58,7 +59,7 @@ class Days360Test extends TestCase
         return require 'tests/data/Calculation/DateTime/DAYS360.php';
     }
 
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerUnhappyDAYS360')]
+    #[DataProvider('providerUnhappyDAYS360')]
     public function testDAYS360UnhappyPath(string $expectedException, mixed ...$args): void
     {
         $arguments = new FormulaArguments(...$args);
@@ -91,7 +92,8 @@ class Days360Test extends TestCase
         self::assertSame(31, Days::between($obj1, $obj2));
     }
 
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerDays360Array')]
+    /** @param mixed[] $expectedResult */
+    #[DataProvider('providerDays360Array')]
     public function testDays360Array(array $expectedResult, string $startDate, string $endDate, ?string $methods): void
     {
         $calculation = Calculation::getInstance();