+++ tests/PhpSpreadsheetTests/Calculation/Functions/DateTime/DaysTest.php
@@ -11,18 +11,19 @@ use PhpOffice\PhpSpreadsheet\Calculation\DateTimeExcel\Days;
 use PhpOffice\PhpSpreadsheet\Calculation\Exception as CalculationException;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
 use PhpOffice\PhpSpreadsheetTests\Calculation\Functions\FormulaArguments;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 class DaysTest extends TestCase
 {
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerDAYS')]
+    #[DataProvider('providerDAYS')]
     public function testDirectCallToDAYS(int|string $expectedResult, int|string $date1, int|string $date2): void
     {
         $result = Days::between($date1, $date2);
         self::assertSame($expectedResult, $result);
     }
 
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerDAYS')]
+    #[DataProvider('providerDAYS')]
     public function testDAYSAsFormula(mixed $expectedResult, mixed ...$args): void
     {
         $arguments = new FormulaArguments(...$args);
@@ -34,7 +35,7 @@ class DaysTest extends TestCase
         self::assertSame($expectedResult, $result);
     }
 
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerDAYS')]
+    #[DataProvider('providerDAYS')]
     public function testDAYSInWorksheet(mixed $expectedResult, mixed ...$args): void
     {
         $arguments = new FormulaArguments(...$args);
@@ -57,7 +58,7 @@ class DaysTest extends TestCase
         return require 'tests/data/Calculation/DateTime/DAYS.php';
     }
 
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerUnhappyDAYS')]
+    #[DataProvider('providerUnhappyDAYS')]
     public function testDAYSUnhappyPath(string $expectedException, mixed ...$args): void
     {
         $arguments = new FormulaArguments(...$args);
@@ -90,7 +91,8 @@ class DaysTest extends TestCase
         self::assertSame(31, Days::between($obj1, $obj2));
     }
 
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerDaysArray')]
+    /** @param mixed[] $expectedResult */
+    #[DataProvider('providerDaysArray')]
     public function testDaysArray(array $expectedResult, string $startDate, string $endDate): void
     {
         $calculation = Calculation::getInstance();