+++ tests/PhpSpreadsheetTests/Calculation/Functions/Financial/DollarDeTest.php
@@ -5,10 +5,11 @@ declare(strict_types=1);
 namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Financial;
 
 use PhpOffice\PhpSpreadsheet\Calculation\Calculation;
+use PHPUnit\Framework\Attributes\DataProvider;
 
 class DollarDeTest extends AllSetupTeardown
 {
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerDOLLARDE')]
+    #[DataProvider('providerDOLLARDE')]
     public function testDOLLARDE(mixed $expectedResult, mixed ...$args): void
     {
         $this->runTestCase('DOLLARDE', $expectedResult, $args);
@@ -19,7 +20,8 @@ class DollarDeTest extends AllSetupTeardown
         return require 'tests/data/Calculation/Financial/DOLLARDE.php';
     }
 
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerDollarDeArray')]
+    /** @param mixed[] $expectedResult */
+    #[DataProvider('providerDollarDeArray')]
     public function testDollarDeArray(array $expectedResult, string $argument1, string $argument2): void
     {
         $calculation = Calculation::getInstance();