+++ tests/PhpSpreadsheetTests/Calculation/Functions/Financial/NPerTest.php
@@ -4,9 +4,12 @@ declare(strict_types=1);
 
 namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Financial;
 
+use PHPUnit\Framework\Attributes\DataProvider;
+
 class NPerTest extends AllSetupTeardown
 {
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerNPER')]
+    /** @param mixed[] $args */
+    #[DataProvider('providerNPER')]
     public function testNPER(mixed $expectedResult, array $args): void
     {
         $this->runTestCase('NPER', $expectedResult, $args);