+++ tests/PhpSpreadsheetTests/Shared/Trend/ExponentialBestFitTest.php
@@ -10,6 +10,9 @@ use PHPUnit\Framework\TestCase;
 class ExponentialBestFitTest extends TestCase
 {
     /**
+     * @param array<mixed> $expectedSlope
+     * @param array<mixed> $expectedIntersect
+     * @param array<mixed> $expectedGoodnessOfFit
      * @param array<float> $yValues
      * @param array<float> $xValues
      */